Move suites into new folder
