+++ src/testing/index.ts
@@ -1,19 +1,19 @@
 import { env } from "process";
 import vscode from "vscode";
 import { instance } from "../instantiate";
-import { ActionSuite } from "./action";
-import { ComponentSuite } from "./components";
-import { ConnectionSuite } from "./connection";
-import { ContentSuite } from "./content";
-import { DebugSuite } from "./debug";
+import { ActionSuite } from "./suites/action";
+import { ComponentSuite } from "./suites/components";
+import { ConnectionSuite } from "./suites/connection";
+import { ContentSuite } from "./suites/content";
+import { DebugSuite } from "./suites/debug";
 import { DeployToolsSuite } from "./deployTools";
-import { EncodingSuite } from "./encoding";
-import { FilterSuite } from "./filter";
-import { ILEErrorSuite } from "./ileErrors";
-import { SearchSuite } from "./search";
-import { StorageSuite } from "./storage";
+import { EncodingSuite } from "./suites/encoding";
+import { FilterSuite } from "./suites/filter";
+import { ILEErrorSuite } from "./suites/ileErrors";
+import { SearchSuite } from "./suites/search";
+import { StorageSuite } from "./suites/storage";
 import { TestSuitesTreeProvider } from "./testCasesTree";
-import { ToolsSuite } from "./tools";
+import { ToolsSuite } from "./suites/tools";
 import { Server } from "../typings";
 import { CoverageCollection, CoverageCollector } from "./coverage";
 import { Tools } from "../api/Tools";
