+++ src/testing/suites/components.ts
@@ -1,9 +1,9 @@
 import assert from "assert";
-import { TestSuite } from ".";
-import { Tools } from "../api/Tools";
-import { GetMemberInfo } from "../components/getMemberInfo";
-import { GetNewLibl } from "../components/getNewLibl";
-import { instance } from "../instantiate";
+import { TestSuite } from "../.";
+import { Tools } from "../../api/Tools";
+import { GetMemberInfo } from "../../components/getMemberInfo";
+import { GetNewLibl } from "../../components/getNewLibl";
+import { instance } from "../../instantiate";
 
 export const ComponentSuite: TestSuite = {
   name: `Component tests`,
