+++ src/testing/suites/content.ts
@@ -4,11 +4,11 @@ import { posix } from "path";
 import tmp from 'tmp';
 import util, { TextDecoder } from 'util';
 import { Uri, workspace } from "vscode";
-import { TestSuite } from ".";
-import { Tools } from "../api/Tools";
-import { getMemberUri } from "../filesystems/qsys/QSysFs";
-import { instance } from "../instantiate";
-import { CommandResult, IBMiObject } from "../typings";
+import { TestSuite } from "../.";
+import { Tools } from "../../api/Tools";
+import { getMemberUri } from "../../filesystems/qsys/QSysFs";
+import { instance } from "../../instantiate";
+import { CommandResult, IBMiObject } from "../../typings";
 
 export const ContentSuite: TestSuite = {
   name: `Content API tests`,
